refactor: duplicated loading in user detail page
- extract ToolbarUserTitleItem into own component
- handle loading state with Loading Skeleton inside ToolbarUserTitleItem
- User Permissiond Detail - add loading state more local to basic data card
