+++ packages/admin/cms-admin/src/userPermissions/user/UserPageToolbar.tsx
@@ -1,47 +1,20 @@
-import { gql, useQuery } from "@apollo/client";
-import { CrudMoreActionsMenu, FillSpace, Loading, StackToolbar, ToolbarActions, ToolbarBackButton, ToolbarTitleItem } from "@comet/admin";
+import { CrudMoreActionsMenu, FillSpace, StackToolbar, ToolbarActions, ToolbarBackButton } from "@comet/admin";
 import { ImpersonateUser, Reset } from "@comet/admin-icons";
-import { styled } from "@mui/material/styles";
 
 import { commonImpersonationMessages } from "../../common/impersonation/commonImpersonationMessages";
 import { ContentScopeIndicator } from "../../contentScope/ContentScopeIndicator";
 import { useCurrentUser, useUserPermissionCheck } from "../hooks/currentUser";
 import { startImpersonation, stopImpersonation } from "../utils/handleImpersonation";
-import { type GQLUserPageQuery, type GQLUserPageQueryVariables } from "./UserPageToolbar.generated";
+import { ToolbarUserTitleItem } from "./userTitleItem/ToolbarUserTitleItem";
 
 export const UserPermissionsUserPageToolbar = ({ userId }: { userId: string }) => {
     const currentUser = useCurrentUser();
     const isAllowed = useUserPermissionCheck();
 
-    const { data, error, loading } = useQuery<GQLUserPageQuery, GQLUserPageQueryVariables>(
-        gql`
-            query UserPage($id: String!) {
-                user: userPermissionsUserById(id: $id) {
-                    name
-                    email
-                }
-            }
-        `,
-        {
-            variables: { id: userId },
-        },
-    );
-
-    if (error) {
-        throw new Error(error.message);
-    }
-
-    if (loading || !data) {
-        return <Loading />;
-    }
-
     return (
         <StackToolbar scopeIndicator={<ContentScopeIndicator global />}>
             <ToolbarBackButton />
-            <ToolbarTitleItem>
-                <TitleText>{data.user.name}</TitleText>
-                <SupportText>{data.user.email}</SupportText>
-            </ToolbarTitleItem>
+            <ToolbarUserTitleItem userId={userId} />
             <FillSpace />
             <ToolbarActions>
                 {isAllowed("impersonation") && (
@@ -66,17 +39,3 @@ export const UserPermissionsUserPageToolbar = ({ userId }: { userId: string }) =
         </StackToolbar>
     );
 };
-
-const TitleText = styled("div")`
-    font-weight: ${({ theme }) => theme.typography.fontWeightRegular};
-    font-size: 18px;
-    line-height: 21px;
-    color: ${({ theme }) => theme.palette.text.primary};
-`;
-
-const SupportText = styled("div")`
-    font-weight: ${({ theme }) => theme.typography.fontWeightRegular};
-    font-size: 14px;
-    line-height: 20px;
-    color: ${({ theme }) => theme.palette.text.secondary};
-`;
