+++ packages/admin/cms-admin/src/userPermissions/user/userTitleItem/ToolbarUserTitleItem.gql.ts
@@ -1,0 +1,11 @@
+import { gql } from "@apollo/client";
+
+export const userTitleItemQuery = gql`
+    query UserTitleItem($id: String!) {
+        user: userPermissionsUserById(id: $id) {
+            id
+            name
+            email
+        }
+    }
+`;
